Project's first edition
this a working version of the project. you can use it as a writer or a reader.
The writer will generate encryption and decryption keys also will encrypt the story which he need to provide.
The reader version will decrypt the story and display it in narration way to the reader
all configurations should be done in the config section in the main.c file